pair program to fix conflicts happened during the merge processs.

+++ db/schema/01_midterm.sql
@@ -8,7 +8,7 @@ DROP TABLE IF EXISTS categories CASCADE;
 CREATE TABLE users (
   id SERIAL PRIMARY KEY NOT NULL,
   name VARCHAR(255) NOT NULL,
-  phone_number INTEGER NOT NULL,
+  phone_number BIGINT NOT NULL,
   password VARCHAR(255) NOT NULL
 );
